adding check if load is dependend to this loop
git-svn-id: https://svn.pnl.gov/svn/hpcgroup/projects/modsim-tools/palm-miami/MIAMI-palm@17735 4c623a4d-1e8e-0410-8a3e-ee3132d44425

+++ src/common/miami_types.h
@@ -68,7 +68,7 @@ struct imm_value_t {
 //ozgurS moving inslvlmap and mem struct to here
 struct memStruct {
    int level;
-   int hitCount;
+   double hitCount;
    double latency;
 };
